--- content/site/skills/bolder.md
@@ -1,40 +1,0 @@
----
-tagline: "Push safe designs toward impact without sliding into chaos."
----
-
-## When to use it
-
-Reach for `/impeccable bolder` when the interface looks like every other interface. Generic sans, medium weights, soft shadows, modest accent color, reasonable spacing, forgettable. The design is not wrong, it is just safe. Use bolder when a project can handle presence and the current state is not bringing any.
-
-Do not use it on dashboards people stare at for hours. Boldness earns its place on marketing pages, hero moments, and content features. Not in operator tools.
-
-## How it works
-
-The skill amplifies four axes without breaking usability:
-
-1. **Scale**: display type gets pushed to clamp(3rem, 6vw, 6rem) or beyond. Headlines that fill the viewport, not hedge it.
-2. **Weight contrast**: light 300 against heavy 800 instead of medium against regular. Real tension, not a shrug.
-3. **Color commitment**: the accent color shows up at full strength, not diluted. Backgrounds can take a stance (ink, accent, cream) instead of all-paper.
-4. **Compositional confidence**: asymmetry, off-grid, pullquotes, hanging punctuation, scale jumps. The layout has a voice.
-
-The skill does not add more. It amplifies what is already there. If the design has three colors, bolder does not add a fourth, it commits harder to the three.
-
-## Try it
-
-```
-/impeccable bolder the landing page hero
-```
-
-Expected changes:
-
-- Hero heading from 3rem to clamp(3.5rem, 7vw, 6.5rem), display font, weight 700
-- Subhead from regular to italic at 1.5rem, pulled 8px left of the heading for optical alignment
-- Background switches from paper to a cream-to-paper gradient, creating a warmer container
-- CTA button fills, drops shadow removed, border radius reduced, hover state inverts colors
-- Supporting image pushed slightly off-grid with a negative top margin, creating asymmetry
-
-## Pitfalls
-
-- **Running it on the wrong page.** Product dashboards, settings, and forms should not be bold. They should be legible. Use `/impeccable layout` or `/impeccable polish` instead.
-- **Confusing bold with loud.** Bold means committed and confident. Loud means shouting. Bolder is the former. If the result feels aggressive, follow up with `/impeccable quieter`.
-- **Pairing it with `/impeccable delight` in the same pass.** Delight works best against a stable visual baseline. Bold first, stabilize, then delight.
