--- content/site/skills/delight.md
@@ -1,42 +1,0 @@
----
-tagline: "Small moments of personality that turn functional into memorable."
----
-
-## When to use it
-
-`/impeccable delight` is for interfaces that work but do not feel like anything. Use it when the core experience is solid and you want to add the small human touches that make people remember it: a considered empty state, a loading message with a point of view, a success animation that feels earned, a microcopy moment that makes someone smile.
-
-It is a finishing skill. Never the first thing you run on a new build.
-
-## How it works
-
-The skill hunts for delight opportunities in the places most designers skip:
-
-1. **Empty states**: instead of "No items yet", something with personality appropriate to the brand.
-2. **Loading and waiting moments**: the best products turn waits into content.
-3. **Success feedback**: a moment of celebration when something worth celebrating happens.
-4. **Microcopy**: button labels, tooltips, error messages, placeholder text. Tiny copy with taste.
-5. **Easter eggs and secondary states**: things users discover that reward paying attention.
-
-The skill reads the brand tone from `PRODUCT.md`. A serious analytics tool gets serious delight (dry, precise, a little clever). A playful consumer app gets more overt personality. It does not force humor where humor is wrong for the audience.
-
-The rule is: every delight moment must still work perfectly if you delete the delight. Nothing depends on the smile.
-
-## Try it
-
-```
-/impeccable delight the first-run experience
-```
-
-Expected additions:
-
-- Empty dashboard replaces "No data yet" with "Your dashboard is quiet. Let's fix that." plus a single-action CTA.
-- Initial sync gets a 3-state loading message that advances: "Finding your accounts... / Pulling the last 30 days... / Making it look good...".
-- First successful action triggers a one-time toast with a tiny celebratory moment. After that, just a quiet checkmark.
-- Help tooltip on the tricky field has a voice that sounds like a person wrote it.
-
-## Pitfalls
-
-- **Forcing humor.** Not every brand is playful. If the brand voice in `PRODUCT.md` is "clinical and precise", delight adds clever restraint, not jokes.
-- **Over-decorating.** One moment of delight is memorable. Twenty becomes noise. The skill is conservative on purpose.
-- **Running delight before polish.** Polish fixes what is wrong. Delight adds what is missing. In that order.
