--- content/site/skills/distill.md
@@ -1,44 +1,0 @@
----
-tagline: "Ruthless subtraction. Strip designs to their essence."
----
-
-## When to use it
-
-`/impeccable distill` removes what should not be there. Competing buttons, redundant information, decorative borders, three fonts where one works, six navigation items where three belong. Use it when an interface feels cluttered, busy, or like it is trying to do too much at once.
-
-Reach for it after `/impeccable critique` flags "cognitive load" or "visual noise", or any time a page has grown by accretion and no one has done the editing.
-
-## How it works
-
-The skill starts from one question: what is the single job this interface is trying to do? Everything that does not help that job is on the chopping block.
-
-It works in two passes:
-
-1. **Assess the complexity sources**. Too many elements, excessive variation, information overload, visual noise, confusing hierarchy, feature creep. Name each one.
-2. **Edit ruthlessly**. Remove what is not essential. Combine what can be combined. Hide what can wait. Consolidate variation into a single treatment. Commit to a single visual language.
-
-The principle: every element on the page has to justify its existence. Fewer obstacles, not fewer features.
-
-## Try it
-
-```
-/impeccable distill this dashboard
-```
-
-Before: four card styles, three button variants, two header treatments, a sidebar with 14 items grouped into 5 sections.
-
-After a `/distill` pass, typical changes:
-
-- Collapse the four card styles into one
-- Pick one button variant, demote the others to text links
-- Unify the headers
-- Group the sidebar into 3 sections, not 5
-- Hide advanced options behind a disclosure
-
-Fewer things. Each one clearer.
-
-## Pitfalls
-
-- **Confusing distill with delete.** Distill removes obstacles. It does not remove features users need. If a user relies on something daily, find a way to keep it quietly, not a way to cut it.
-- **Running it too early.** If the feature is still growing, distilling it now means distilling the same thing again next week. Wait until the shape is stable.
-- **Expecting it to replace hierarchy work.** Sometimes the right fix is not removing things, it is arranging them. Reach for `/impeccable layout` when the problem is layout, not quantity.
