+++ site/data/anti-patterns-catalog.js
